STSX commit
fabSuiteCom 
Lot number missing problem with FabSuiteSave operation
Clean up reported XML for easier identification of XML formatting
Loads not reporting as loaded in FabSuite, so reversal of Load/Delete
Plate width cutting incorrectly assumed as always 48"
Incorrectly capturing model part type in XML
Relations added for import_job_to_table, job_to_idlabel,
job_to_raw_label, customer address and billing indications
jobsts.globals.js
Periodic messaging change causes long pause as well as does not reappear
if called after closing
Changes for shipping and billing customer addresses
Embedded comments interfering with object construction, removed
Piecemark assembly using assembly to indicate major piecemark
Erroneous use of customer barcode settings instead of established
barcode settings forced reduction of barcode to 7 characters
cut cutlist issues incorrect
major piecemark not using major assembly when no minor indicated
Admin users lost ability to find piecemarks and get status, restored
cutover inventory, sheets, piecemarks (major), 
cutover quit capability to abandon further cutover operations
jobsts.jobs.js
stop import or continue if close the message box
piecemark subassembly creation
Always show total count and finished count on import message
Add flush clients after import
If aborted import or import underway, verify with user that they want to
continue
Adjust label count sent to print display message
Backup message prior to import
Catch operation was hiding import errors, removed
Printer label selections were not being honored
Adjust inventory item weight according to cut length, since returned
items are not mirrors of EPM inventory, more of a reality check
inventory location updates
all cutover data has logic flag = 1 that will be later removed
If close of message window, then reopen again
Job name not saved during import
cutover loads,sheets, customers, employees, jobs, piecemarks as above
jobsts.prefs.js
if useServer for printing labels, was ignoring labelase choice, so
modded code to permit labelase printing either under server or not
labelase demands and additional column that is the template path, which
bartender does not, so  needed a bit of rewrite to NOT use bartender
data format
label printing did not require padding of data, so field lengths were
too long and disappeared during print off label
jobsts.printer.js
lot number not showing in available print fields
kiss_import.frm
import quantities were off that did not show correct quantity already in
STSX
color changes and textual changes during import to visually see items
that needed attention
kiss_import.fm. messagesWarnings.js to execute continue actions after
the message window is closed
kiss_import.kiss.js
sheet numbers not chosen when filtered
lots and sequences not filtering correctly during import
properly establish parent if no minor is indicated
remove extra import files
table inventory add disposition for piece
table jobs job_status added for import status
table piecemarks added is_assembly
form barcode_idlabel.js form.barcode_print interfering with each other
requiring a different relation for the views
form gen_table_order.js missing some i18n entries indicating more
readable format for the column ordering selection
form jobs_general.js i18n the initial <Empty> entry
form loads_criteria as per PP leave hide empty columns unchecked
forms MWBase.js add flush clients button
form sts_nav_import.js remove KISS Import button if web client in use
STSmobile form rf_mobile_view.frm labelled name incorrectly causing
screen label overwrite

+++ STS3/forms/kiss_import.js
@@ -214,6 +214,7 @@ function fileReceipt(file){
  */
 function getKissFile(event){
 	if (globals.deBounce(event)){return}
+	scopes.kiss.removeExtraKissFiles(event);
 	controller.enabled = false;
 	randFile = 'Import_'+vJobNumber+'_'+new Date().getTime()+'.kss';
 	scopes.jobs.importJob.guidOrder = 1;
@@ -258,7 +259,6 @@ function getKissFile(event){
 			<IncludeLotNumbers>1</IncludeLotNumbers>\n\
 			</ExportJob>\n\
 			</FabSuiteXMLRequest>';
-		//application.output(request);
 		if (vSeqNumber && vSeqNumber.search(',') == -1){
 			request = request.replace('SEQ','<Sequence>'+vSeqNumber+'</Sequence>');
 		} else {
@@ -545,20 +545,25 @@ function onDataChangeLotSeq(oldValue, newValue, event) {
 function removeJobImportData(jobNumber){
 	var date = new Date().getTime() - 1000*60*60*24;
 	var delDate = new Date(date);
-	delDate = utils.dateFormat(delDate, 'yyyy-MM-dd')
+	//delDate = utils.dateFormat(delDate, 'yyyy-MM-dd')
 	/** @type {QBSelect<db:/stsservoy/import_table>} */
 	var q = databaseManager.createSelect('db:/stsservoy/import_table');
-	q.where.add(q.or
-		.add(q.columns.job_number.eq(jobNumber))
-		.add(q.columns.creation_date.lt(delDate))
-	);
+	q.where.add(q.columns.job_number.eq(jobNumber));
 	q.where.add(q.columns.tenant_uuid.eq(globals.session.tenant_uuid));
 	q.result.add(q.columns.import_table_id);
+	/** @type {JSFoundSet<db:/stsservoy/import_table>} */
+	var Q = null;
+	Q = databaseManager.getFoundSet(q);
+	databaseManager.saveData(Q);
 	var jobId = null;
-	var Q = databaseManager.getFoundSet(q);
-	if (Q.getSize() > 0){
+	if (application.isInDeveloper()){application.output('size of import table records '+Q.getSize())}
+	if (Q && Q.getSize() > 0){
 		var rec = Q.getRecord(1);
 		jobId = rec.job_uuid;
+		Q.loadAllRecords(); var idx = Q.getSize();
+		while (rec = Q.getRecord(idx++)){
+			idx = Q.getSize()+1;
+		}
 		Q.deleteAllRecords();
 	}
 	if (application.isInDeveloper()){application.output('size of import table records '+Q.getSize())}
